fix build

+++ src/server/shared/Cryptography/ARC4.h
@@ -26,7 +26,7 @@
 class ARC4
 {
     public:
-        ARC4(uint8 len);
+        ARC4(uint32 len);
         ARC4(uint8* seed, uint32 len);
         ~ARC4();
         void Init(uint8* seed);
